+++ src/main/java/prismatic/shards/stellarity/registry/feature/FreezeWaterFeature.java
@@ -6,7 +6,6 @@ import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
 import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
-import prismatic.shards.stellarity.Stellarity;
 
 import static prismatic.shards.stellarity.util.ValueUtil.from;